add loggers and change price to bigdecimal type

+++ admin-service/src/test/resources/initdb.sql
@@ -5,7 +5,7 @@ create table good
     id varchar primary key,
     type varchar(255),
     manufacturer varchar(255),
-    price int,
+    price numeric,
     name varchar(255),
     release_date timestamp
 );
